backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
aiofiles==23.2.1
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0
anthropic==0.26.1
openai==1.30.1
numpy==1.26.4
scikit-learn==1.4.2
sentence-transformers==2.7.0
aiosmtplib==3.0.1
email-validator==2.1.1
celery==5.4.0
redis==5.0.4
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.2
faker==24.14.0
